Fix preparing of mocked method names
Because of preparing method names as indexed arrays and using the
function "array_replace_recursive", different lists of methods overwrote
the items of each other.

To prevent overwritting of method names of duplicating of them the
preparing of method name lists was changed.
Now that lists of method names are the associative arrays.
The keys of that arrays are represented by the method names.